requirements.txt
pip==21.1.1
flask==1.1.2
gunicorn==20.1.0
requests==2.27.1
telegram-send==0.25
pytz==2021.1
google-cloud-bigquery==2.16.1
google-cloud-tasks==2.2.0
google-cloud-secret-manager==2.4.0
google-cloud-datastore==2.4.0
ta==0.8.0
pandas
pandas-gbq
pyarrow
numpy
bson
redis
matplotlib